Refactor `isLiteral`
